requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
pydantic==2.8.2